requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langchain-text-splitters
langgraph
langgraph-prebuilt
chromadb
ddgs
beautifulsoup4
requests
pdfplumber
fastapi
uvicorn
python-dotenv
psycopg2-binary
bcrypt
openpyxl
python-docx
openai-whisper
sounddevice
scipy
numpy
pandas